Refactoring & added a Clock Task.
