+++ src/graphic/drawing.h
@@ -9,6 +9,6 @@
 // Declare any function prototypes here
 
 void drawLineThickness(LGFX_Sprite *sprite, int x1, int x2, int y1, int y2, int color);
-void drawGPSInfo();
+void drawGPSInfo(void *pvParameters);
 
 #endif // DRAWING_H
