+++ src/graphic/task.h
@@ -1,0 +1,12 @@
+#ifndef TASK_H
+#define TASK_H
+
+#include "globalVariables.h"
+
+#include "device/gps.h"
+
+// Define task function prototypes
+void initGPSTask();
+void initClockTask();
+
+#endif // TASK_H
